+++ lib/constants.dart
@@ -1,0 +1,25 @@
+class InternalConstants {
+  /// The string constant for DNS, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zDNS = '6ba7b810-9dad-11d1-80b4-00c04fd430c8';
+
+  /// The string constant for URL, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zURL = '6ba7b811-9dad-11d1-80b4-00c04fd430c8';
+
+  /// The string constant for OID, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zOID = '6ba7b812-9dad-11d1-80b4-00c04fd430c8';
+
+  /// The string constant for X500, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zX500 = '6ba7b814-9dad-11d1-80b4-00c04fd430c8';
+
+  /// The string constant for NIL, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zNIL = '00000000-0000-0000-0000-000000000000';
+
+  /// The string constant for MAX, primarily for internal use, do not use,
+  /// no guarantees this won't be renamed or removed.
+  static const zMAX = 'ffffffff-ffff-ffff-ffff-ffffffffffff';
+}